HackerNews.Connector.Python/requirements.txt
curiosity==36441
hackernews-python==0.3.2